requirements.txt
fastapi
uvicorn
python-multipart
jinja2
groq
elevenlabs
python-dotenv
websockets
